ble_mqtt/requirements.txt
bleak
requests
getmac
paho-mqtt